pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deltapd"
version = "0.2.0"
description = "UHF Partial-Discharge signal processing and Δt tracking framework"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "scipy>=1.10",
  "PyWavelets>=1.4",
  "pandas>=2.0",
  "h5py>=3.0",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "black>=24.0",
  "ruff>=0.4",
  "pytest>=8.0",
]

[project.scripts]
deltapd = "deltapd.__main__:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
